Rename ::normalize to ::normalize_value_for
The `::normalize` method accepts an attribute name and a value, and type
casts the value, applying any declared normalizations for the attribute.
Because of the way type casting works for most types, the type of the
given value can be very different than the attribute type.  This allows
for potential confusion between the `::normalize` method and the
`::normalizes` method.  For example:

  ```ruby
  class Post < ActiveRecord::Base
    normalize :title, with: -> { _1.titlecase }
  end
  ```

In the above code, the user meant to call `::normalizes`, but called
`::normalize` instead.  The `:with` kwarg is treated as a `Hash` value
which is then type cast to a `String`.  The type cast silently succeeds,
so the user does not realize why their normalization is not applied.

To prevent such confusion, this commit renames `::normalize` to
`::normalize_value_for`.