fix(autosave-association): search for records that need to be autosaved or that need validation
In order to solve the problem of issue #46509 where when there is an attribute that must be only integer it did not generate an error message when it was used through accepts_nested_attributes.

The main cause is that, in order to validate the accepts_nested_attributes item, it is always sought if any records hear changes, but since the interger validation makes a type_cast, a change is never detected when only the fractional is changed... 4 -> 4.1

With that, a new method was added that looks to see if there is any change between the before_type_cast value and the current value. And with that search them in the autosave_association