Add missing requires in generated plugin bin/rails
Prior to this commit, when generating a plugin with the `--full` option
and a framework skip option (e.g. `--skip-action-cable`),
`require "action_mailbox/engine"` and `require "action_text/engine"`
would be missing from the generated `bin/rails` file.

This commit adds the missing `require` statements.

This is a manual backport of the fix from #47015, minus the refactoring.