Generate `Dockerfile` to exclude `test` Bundle group
The generated `Dockerfile` excludes the `development` group, but does
not exclude the `test` group. Since it's `production`-focused, it should
exclude both.

According to the [Bundler Configuration][] documentation, you can
specify a `--without` flag or `BUNDLE_WITHOUT` environment variable with
the group names to omit from bundling, separated by a `:`

> without (BUNDLE_WITHOUT): A :-separated list of groups whose gems
> bundler should not install.

[Bundler Configuration]: https://bundler.io/v2.4/man/bundle-config.1.html#CONFIGURATION-KEYS