Fix test:units and test:functionals for a new application
When creating a new application the `test/unit` and `test/functional`
directories are no longer created. Running `bin/rails test:units` or
`bin/rails test:functionals` tries to run tests in these directories but
will fail when any of the directories are missing.

By using a glob pattern for the tests the task will no longer fail when
run.