Adds support for `UPDATE ... FROM ...` syntax to Arel
This is useful to update one table based on the contents of another table, a common table expression, or a subquery. `UPDATE ... FROM ...` is supported in PostgreSQL and Sqlite3.
While MySQL doesn't explicity suppor the syntax, there's potentially a way to replicate this behavior by including the from relation in the list of tables to be updated.

Very similar work was also done by @quentindemetz and @lazaronixon

Co-authored-by: Nixon <lazaronixon@hotmail.com>
Co-authored-by: Quentin de Metz <quentin@de.me.tz>