Simplify asset manifest example
There's no need to paste the entire file, and replace the actual values with placeholders.
This saves a ton of space on space on the page and focuses on the important details.

Also updated the code block to JSON to avoid parsing the comment as Ruby,
this also means keys and values get highlighted appropriately.