Fix preview variants in ActiveStorage proxy mode
Fixes #47071

Before this commit, when using ActiveStorage proxy mode, it was
impossible to apply variations to the generated preview of a
video or a PDF. This is because `ActiveStorage::Preview#image` (which
is called in the `ActiveStorage::Representations::ProxyController`)
always returned the blob's preview image without applying any
variations.

This commit fixes the issue by changing `ActiveStorage::Preview#image`
so that it takes into account the variations.