Get the information from the right database
Previous this tag would get the information only from ActiveRecord::Base
connection, which in a multi-database application would be wrong.

This get the information from the connection that is going to be used to
execute the query.

We also had to change the NullPool to make this work.

It is only required because when checking out a connection form the pool,
we first create the object, check if the database version satisfies the
requirements, and then we assign the pool to the connection.

When we are checking the database version, we are doing a query, so we
try to annotate the query with the tags, but the pool isn't assigned yet,
so `pool.db_config` would return an exception.

This is because the NullPool doesn't satisfy the ConnectionPool
interface, which we should fix.

As the only queries that happen without a pool are very likely internal
queries we are retuning an object that respond to anything, without
caring if that object have the same interface of the real configs.