Remove unneeded require in plugin application.rb
This require was added in cfcea1d53ae5ce38a7cbeb41e05958dc009988b0, but
at that time the generated Gemfile did not resolve using the gemspec.
gemspec was added to the plugin's generated Gemfile later in
a74e4736f95befa7a22c208019bf11a155ff7543. Since then, the
Bundler.require in the plugin's generated config/application.rb will
require the plugin and the extra require is unneeded.

The replaced test was added in adfa417fbbf6f670c3d10ed2c32c58bc1dc92c2e
when the plugin generator's application.rb was replaced with the app
generator's. Since the test is only concerned with the file content and
not whether that content is useful, the test was replaced with an
attempt to run the plugin's test, since this would verify that the
plugin can be loaded within the dummy app.

The plugin helpers are added because the generated Gemfile is now used
for resolution instead of Rails'. This makes the tests more accurate
because real plugins will be resolving against their own Gemfiles.
However, this also leads to a few issues solved by the new helpers:
- gemspecs are generated with TODOs that have to be fixed
- commands run in the plugin must have their Bundler environment reset
because processes spawned in Rails' tests inherit BUNDLE_GEMFILE

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>