Avoid assuming that save is instant
When unlucky, it can take a few milliseconds, and if particularly
unlucky that plus precision truncation can push us outside the previous
1s window.

Instead, we'll accept a +/- 1s buffer around the whole time that was
spent doing the save. This test is only about ensuring that some vaguely
CURRENT_TIMESTAMP-shaped value (i.e. the column's default) was stored,
and isn't actually so concerned about the specifics, anyway.

This test is still technically at risk of failure if the test-running
system's time disagrees with the database server's, or if the system
clock is adjusted mid-run... but those still seem acceptably ignorable;
I'm just reacting to a flaky failure in CI.