Regenerate Active Storage dummy app
Generated using the following script and manually reviewed after:
skipkayhil/rails-bin@cbad7eddcb439c44e5ec28ade3336245917542fe

The active_storage configuration had to be moved to the dummy app
because of the changed eager_load config. Now, the whole app is eagerly
loaded when the app's environment file is required which means that
changing configuration in the test_helper is too late.