Apply suggestions from Action Controller Overview guide
There were some changes from #46972 that I forgot to make here as well.

I'm intentionally duplicating this information because it's a new feature and I want users to be able to find it where they are.
Whether they are reading the guides online or the API reference somewhere, like in their editor.