Delete test that tests nothing
I deleted the `establish_connection` in the middle of the test and it
still passed.

This test has no asserts and I can't figure out what I was trying to
test when I added it in 1ee4a8812fc. We have other tests testing
advisory locks, so remove this test.