Clean up async test
Due to the ensure on the `setup` method, we weren't actually
testing against a config with `min_threads` set to 0 and `max_threads`
set to 10. The `ensure` was setting the connections back to their
original state before the test ran. In this case the test flow was:

* call setup (setup connections)
* call ensure from setup (reset connections)
* run test
* call teardown

We want the flow to be:

* call setup (setup connections)
* run test
* call teardown (reset connections)

I also simplified the configuration setup, we can get it from the
existing connection, merge changes in, and then establish a connection.
I added a test asserting the configs because without this the file
passes since the min/max threads are never exercised.