Make ForkTracker.check! a noop on Ruby 3.1+
Fix: https://github.com/rails/rails/issues/47418

We should trust the `Process._fork` callback, as it can't really
be bypassed except throught `Process.daemonize` but I don't think
it's a problem.

We initially kept the PID check even if the callback was present,
but it turns out that on modern GLIBC the getpid() syscall isn't
cached and it causes a significant overhead.