Ensure auto_increment declaration is preserved when copying SQLite tables
Fixes #47400.
Previously, when copying a table as a result of an ALTER TABLE
statement, the AUTOINCREMENT declaration on the PK column was dictated
by the absence of a `:default` option on that column. However,
we were always passing `default` when copying the columns to the new table,
regardless of whether the column originally had a default. This led to
the AUTOINCREMENT declaration being lost.

This commit fixes the issue by tracking a column's auto_increment value,
and only passing the `default` option in `#copy_table` when the column
does not have the `auto_increment` option set.

Co-authored-by: fatkodima <fatkodima123@gmail.com>