Add response fields to request.action_dispatch
An issue was opened about Rails logging a 200 status when the actual
status returned to the client was a 304 (because it was changed by
Rack::ConditionalGet). This is due to requests being logged from the
process_action.action_controller instrumentation, which is at the very
bottom of the Rack middleware stack.

To help fix this issue, request.action_dispatch has been improved
to include response status and headers. This makes the instrumentation
more usable for request logging, as the payload now contains references
to almost all of the data provided in process_action.action_controller.
In addition, its much higher position in the middleware stack means that
it will return a "more accurate" status than  is available in
process_action.action_controller.