Allow 3-tier shared config in database.yml
Previously, the shared config hash would be merged with all database
configurations defined in database.yml.

This commit makes the shared hash 3-tier aware so that it can be used
to share configuration across environments but only with matching
configuration names.

For example:

```yaml
shared:
  one:
    migrations_path: "db/one"
  two:
    migrations_path: "db/two"

development:
  one:
    adapter: sqlite3
  two:
    adapter: sqlite3

production:
  one:
    adapter: mysql2
  two:
    adapter: mysql2
```

Will now properly set the migration_path for the one and two
configurations in both development and production.