Fix `rails new --dev APP_PATH` command crashing
ref: https://github.com/rails/rails/issues/47435#issuecomment-1436530001

- `rails new APP_PATH --dev` works fine.
- `rails new --dev APP_PATH` does not work.

For most other flags, putting the APP_PATH before or after the flags has the same effect. It's just the prelease ones that have issues, because we have to rengerate the command so as to call `bundle exec rails new ...` a second time.

Prior to this PR, the command that was being called was `bundle exec rails new APP_PATH APP_PATH`, and that doesn't work. Now it will get called as `bundle exec rails new APP_PATH --dev APP_PATH` which is fine.

I haven't been able to find a way to write an automated test for this, but you can use the replication steps in https://github.com/rails/rails/issues/47435 to verify that the issue is fixed.

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>