bring Rails 7.1 template to parity with dockerfile-rails
This covers all of the support in dockerfile-rails that applies to
the "rails new" scenario.  It is being offered in the spirit of not
having an unintentional differences between these two approaches.

Intentional differences are fine.  Feel free to cherry pick.  Feel
free to push changes directly into this pull request.

One notable exception: the code mentioned by the following comment
is *NOT* included in this commit as it requires more analysis:

https://github.com/rails/rails/pull/47460#issuecomment-1440810111

cc: @byroot, @dhh, @zzak