Only require irb if console is unconfigured
Previously, irb would be required by the console whether or not another
console was configured. While this is not a huge issue at the moment, it
will become an issue if irb is removed as a default gem (which has been
proposed for Ruby 3.3).

This commit changes the console command to only require irb if it was
not previously configured, preventing potential issues with it not being
included in an application's dependencies.

There is currently an open discussion about whether irb should be added
to generated Gemfiles. This change is beneficial whether or not that PR
is merged:
- if irb is added to Gemfiles, then this change will prevent errors
  running the rails console for apps that use other consoles
- if irb is added to railties dependencies, then this change will
  result in less libraries required for apps that use other consoles

The tests needed to be updated because of the ordering of the assertion
parameters. Since irb is now only required when the command is run, the
constant can not be referenced first.