Fix parsing ruby version for Dockerfile packages
Previously, the version regular expressions never matched
Gem.ruby_version because Gem::Version is not a string. In addition,
ruby_version was undefined and this was not caught before because the
case statement never evaluated those branches.

This commit fixes both of those issues, so that ruby versions like 2.7.0
will correctly install "python" instead of "python-is-python3"