bring Rails 7.1 template to parity with dockerfile-rails - take 2
This covers all of the support in dockerfile-rails that applies to
the "rails new" scenario.  It is being offered in the spirit of not
having an unintentional differences between these two approaches.

Intentional differences are fine.  Feel free to cherry pick.  Feel
free to push changes directly into this pull request.

As before, this does not cover yarn as it requires more analysis:
* https://github.com/rails/rails/issues/47416
* https://github.com/rails/rails/pull/47460#issuecomment-1440810111

This pull request supercedes https://github.com/rails/rails/pull/47465