Use squiggly heredoc to remove extra spaces after `DEPRECATION WARNING:`
- Without this commit

```ruby
DEPRECATION WARNING:                 Passing the coder as positional argument is deprecated and will be remove in Rails 7.2.

                Please pass the coder as a keyword argument:

                  serialize :foo, coder: JSON
 (called from <class:User> at rep.rb:20)
DEPRECATION WARNING:                 Passing the class as positional argument is deprecated and will be remove in Rails 7.2.

                Please pass the class as a keyword argument:

                  serialize :bar, type: Hash
 (called from <class:User> at rep.rb:21)

```

- With this commit

```ruby
DEPRECATION WARNING: Passing the coder as positional argument is deprecated and will be remove in Rails 7.2.

Please pass the coder as a keyword argument:

  serialize :foo, coder: JSON
 (called from <class:User> at rep.rb:20)
DEPRECATION WARNING: Passing the class as positional argument is deprecated and will be remove in Rails 7.2.

Please pass the class as a keyword argument:

  serialize :bar, type: Hash
 (called from <class:User> at rep.rb:21)
```