Ensure app generator works on Ruby 2.7 rubygems
Previously, this method compared Gem::Versions with Strings, which isn't
supported until Rubygems 3.3.6.

Now, it correctly only compares Gem::Versions with Gem::Versions.

An alternative approach could be to use Gem::Version.new(RUBY_VERSION),
however there doesn't appear to be a benefit in this case and the
downside is that testing becomes much harder.