Add test:arel task to run only Arel tests
Previously, all of the Arel tests would be run with every database
adapter. This is not necessarily a problem, but these tests end up
getting run along with every adapter's tests in CI even though none of
them require a database.

This commit starts the process of testing Arel separately from database
adapters by adding a new task to run Arel's tests. The next step will
be to add the test:arel task to Buildkite, and then the Arel tests can
be filtered from the list of adapter tests so that they are only run
once.