Improve tests for read_attribute when attribute is missing
A MissingAttributeError is raised when a valid column isn't selected in
the query. The assertions for this behaviour are a bit unclear.

It is currently unclear that `:extendedWarranty` is a valid attribute
for the Computer model, because of its uncommon camel-casing. It raises
an error because the column isn't selected in the query.
By adding `:extendedWarranty` to the select statement, we can see that
only columns that aren't selected, for example `:developer`, raise the error.

The test currently also has assertions for writing the attribute.
These assertions are moved to the write_attribute tests.