Make SchemaCache#connection thread-safe
AbstractAdapater#schema_cache assigns SchemaCache#connection=self upon each
use. When schema_cache is called in parallel and has to call the database (as
is particularly common right after app launch), this assignment is subject to
a race condition and multiple threads can end up using the same connection.

This changes SchemaCache#connection to a thread-local variable to avoid this.